Updated REAMDE and started working on get_next_line (will have to restart 'cause I don't understand what I did)

+++ ft_printf2/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: saeby <saeby@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/31 15:49:21 by saeby             #+#    #+#             */
-/*   Updated: 2022/11/02 07:30:19 by saeby            ###   ########.fr       */
+/*   Updated: 2022/11/02 17:16:39 by saeby            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 # include <stdlib.h>
 # include <string.h>
 # include <stdarg.h>
+# include <limits.h>
 # include <stdio.h>
 
 typedef struct s_flags
@@ -38,7 +39,7 @@ int		ft_print_d(int nbr, t_flags *flags);
 int		ft_print_u(unsigned int nbr, t_flags *flags);
 int		ft_print_x(unsigned int nbr, int format, t_flags *flags);
 int		ft_print_s(char *s, t_flags *flags);
-int		ft_print_p(int adr, t_flags *flags);
+int		ft_print_p(long long adr, t_flags *flags);
 int		ft_strlen(char *s);
 
 int		ft_putchar(char c);
